- TranDanh commit ObjectStage_1. Con thieu Boss + bridge

+++ SolutionContraGame/ContraGame/CObjectDx9.h
@@ -25,8 +25,8 @@ protected:
 	CSpriteDx9*		m_Sprite;
 	eObjectState	m_ObjectState;
 	ETypeObject		m_TypeObject;
-	float			m_TimeChangeState;
-	float			m_TimeChangeDirectAttack;
+	int			m_TimeChangeState;
+	int			m_TimeChangeDirectAttack;
 	int				m_AttackCounter;
 	static int		m_identifyNumber;
 	int				m_ObjectIdentifyNumber;
